+++ src/main/java/de/tomalbrc/cameraobscura/util/Constants.java
@@ -2,6 +2,5 @@ package de.tomalbrc.cameraobscura.util;
 
 public class Constants {
     public static final String DYNAMIC_PLAYER_TEXTURE = "dyn.p"; // player models and player heads
-    public static final String DYNAMIC_SIGN_TEXTURE = "dyn.s";
-    public static final String DYNAMIC_MAP_TEXTURE = "dyn.m";
+    public static final String DYNAMIC_DISPLAY_TEXTURE = "dyn.d";
 }