+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.4] - 2025-08-20
 ### Added
 - Modded block texture support
+- Computer display rendering
 
 ## [1.2.3] - 2025-08-19
 ### Fixed